Provide options to filter out pseudo dependencies (#1)
Initial implementation of the composer plugin. This allows a composer project to specify which packages should be ignored during composer install and update.

We mainly need this to ignore the shopware dependency inside shopware 6 plugins. We also need to ignore dependencies to other Shopware 6 plugins on which me might rely.